charlie-kirk-project/requirements.txt
opencv-python
mediapipe==0.10.14
pygame
numpy
Pillow
pystray
pyinstaller
